+++ src/main/java/com/svalero/tiendaAlimentos/domain/RegistrosConsumoAlimentos.java
@@ -11,7 +11,7 @@ public class RegistrosConsumoAlimentos {
     private int id;
     private int usuario_id;
     private int alimento_id;
-    private double cantidad_consumida;
+    private float cantidad_consumida;
     private Date fecha_consumo;
 
 }
