+++ src/main/java/com/svalero/tiendaAlimentos/domain/Usuarios.java
@@ -20,8 +20,8 @@ public class Usuarios {
     private String nivel_actividad;
     private String objetivo_salud;
     private String preferencias_alimenticias;
-    private double altura;
-    private double peso;
+    private float altura;
+    private float peso;
     private Date fecha_nacimiento;
     private String rol;
 }
